+++ src/layouters/threeDiffToBaseMergedRightLayouter.ts
@@ -1,9 +1,9 @@
-import { DiffedURIs } from "../diffedURIs";
-import { Monitor } from "../monitor";
-import { TemporarySettingsManager } from "../temporarySettingsManager";
-import { VSCodeConfigurator } from "../vSCodeConfigurator";
 import { Zoom } from "../zoom";
-import { DiffLayouter, DiffLayouterFactory } from "./diffLayouter";
+import {
+  DiffLayouter,
+  DiffLayouterFactory,
+  DiffLayouterFactoryParameters,
+} from "./diffLayouter";
 import {
   diffEditorSymbol,
   GroupOrientation,
@@ -14,16 +14,12 @@ export class ThreeDiffToBaseMergedRightLayouterFactory
   implements DiffLayouterFactory {
   public readonly settingValue = "3DiffToBaseMergedRight";
 
-  public create(
-    monitor: Monitor,
-    temporarySettingsManager: TemporarySettingsManager,
-    diffedURIs: DiffedURIs,
-    vSCodeConfigurator: VSCodeConfigurator
-  ): DiffLayouter {
-    return new SplitDiffLayouter(
-      monitor,
-      diffedURIs,
-      (diffedURIs, zoom) => {
+  public create(parameters: DiffLayouterFactoryParameters): DiffLayouter {
+    return new SplitDiffLayouter({
+      ...parameters,
+      supportedZooms: [Zoom.top, Zoom.bottom, Zoom.left, Zoom.right],
+      mappedIntervalRelativeSize: 1 / 3,
+      createLayoutDescription: (diffedURIs, zoom) => {
         let topSize = 0.5;
         let leftSize = 0.5;
         switch (zoom) {
@@ -83,9 +79,6 @@ export class ThreeDiffToBaseMergedRightLayouterFactory
           ],
         };
       },
-      temporarySettingsManager,
-      vSCodeConfigurator,
-      1 / 3
-    );
+    });
   }
 }